Refactor
Added a User Intake that creates a user using LiteDB stores a random generated Guid and saves all the users intake information. The Guid is used to load the users information. This is not meant to be a multi-user application so this just loads the single user(first user) if the Guid doesnt match or doesnt exist then it runs the User intake if one exists and matches it loads the home page. Additionally started on the complete refactor of the MyBodyPage currently only has the users Name, Gender, DOB, and Height loading and editable on the page.