Goal Update 1
Updated the Goal data in the DB model to store TimeFrame as TimeframeAmount and TimeframeUnit. This will be used for calculating the time remaining on the goal. Changed the UserIntake to instead take those two variable via input box and combo box. Then added functionality to be able to edit and delete the goals in case a mistake is made during user intake.