Workout Plan Update
Created a new system for the workout plan that works much better for users. Instead of sticking to a Sunday-Saturday method for workouts we instead allow the user to select what length of workout cycle they want to use 7,10,or 14 days. weeks will still be used for progress but for workout plans and streaks this cycle plan works better and gives the users more flexibility in how they plan their workouts. also added the rest day selection. instead of live updates to the workout plan we use a edit/save functionality. only thing that needs to be added is the ability to delete an exercise in the DailyWorkoutPlanner component.